Support #include for cpp script files

+++ sandbox/scripts/test_script.h
@@ -3,7 +3,7 @@
 DEFINE_SCRIPT_INTERFACE(TestScript,
 struct TestScript
 {
-    virtual int foo(int a, int b) = 0;
+    virtual double foo(int a, double b) = 0;
     virtual int bar(int a) = 0;
 };
 )
